requirements.txt
dash
flask
pandas
plotly
couchdb
faker
python-dotenv
